rename: 쪽지시험 관련 응답 타입을 컨벤션에 맞는 위치로 이동 (#78)
